finish order create

+++ app/models/order.go
@@ -21,7 +21,7 @@ const (
 type Order struct {
 	model.MongoModel
 	OrderNo      string                 `json:"order_no" bson:"order_no" name:"订单号"`
-	ItemCount    int                    `json:"item_count" bson:"item_count" name:"订单商品数量"`
+	ItemCount    int64                    `json:"item_count" bson:"item_count" name:"订单商品数量"`
 	OrderAmount  uint64                 `json:"order_amount" bson:"order_amount" name:"订单金额,单位分"`
 	ActualAmount uint64                 `json:"actual_amount" bson:"actual_amount" name:"实付金额,单位分"`
 	OrderItems   []*OrderItem           `json:"order_items" bson:"order_items" name:"订单详情"`
